removed MainAuthorizationsView
